third_party/blink/web_tests/platform/generic/http/tests/websocket/fragmented-binary-frames-expected.txt

WebSocket: Receive fragmented binary messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS receivedMessages.length is 4
Checking message #0.
PASS responseType is "[object Blob]"
PASS actualArray.length is 13
PASS Passed: Message #0.
Checking message #1.
PASS responseType is "[object Blob]"
PASS actualArray.length is 13
PASS Passed: Message #1.
Checking message #2.
PASS responseType is "[object Blob]"
PASS actualArray.length is 0
PASS Passed: Message #2.
Checking message #3.
PASS responseType is "[object Blob]"
PASS actualArray.length is 256
PASS Passed: Message #3.
PASS successfullyParsed is true

TEST COMPLETE
